tests/pyproject.toml
[tool.poetry]
name = "tests"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pytest = "^7.1.2"
psutil = "^5.9.1"
pyotp = "^2.6.0"
paramiko = "^2.11.0"
Flask = "^2.2.1"
requests = "^2.28.1"
flask-sock = "^0.5.2"
websocket-client = "^1.3.3"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.2"
pytest = "^7.1.2"
black = "^22.10.0"

[tool.pytest.ini_options]
minversion = "6.0"
filterwarnings = ["ignore::urllib3.exceptions.InsecureRequestWarning"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
